pyproject.toml
[tool.poetry]
name = "text-tagging-model"
version = "0.1.3"
description = "Here we collected some online and offline models for text tagging."
authors = [
    "Pavel Kochkin <kochkin27@gmail.com>",
    "Shamil Nurkaev <nurkaievs1204@gmail.com>",
    "Denis <denison03@mail.ru>",
    "Stephan Olizko <stepa.olizko@yandex.ru>"
]
readme = "README.md"
packages = [{include = "text_tagging_model"}]

[tool.poetry.dependencies]
python = "^3.9"
rouge = "^1.0.1"
scikit-learn = "^1.4.1.post1"
pymorphy2 = "^0.9.1"
nltk = "^3.8.1"
rake-nltk = "^1.0.6"
coloredlogs = "^15.0.1"
tqdm = "^4.66.2"
fasttext-wheel = "^0.9.2"
transformers = "^4.39.3"
sentencepiece = "^0.2.0"
sentence-transformers = "^2.6.1"
numpy = "^1.24.4"


[tool.poetry.group.dev.dependencies]
pandas = "~2.0.3"
pylint = "^2.17.4"
flake8 = "^6.0.0"
mypy = "^1.2.0"
pytest = "^7.3.1"
ipykernel = "^6.29.4"
isort = "^5.13.2"
pre-commit = "^3.7.0"
black = "^24.3.0"
coverage = "^7.4.4"
torch = "^2.2.2"


[tool.pytest.ini_options]
pythonpath = [
  ".", "text_tagging_model",
]
testpaths = ["./tests/"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[tool.coverage.run]
branch = true
source = ["text_tagging_model"]
command_line = "-m pytest"

[tool.coverage.report]
show_missing = true

# [tool.coverage.xml]
# output = "coverage.xml"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
